set default margin to 10% from current price on coin init

+++ src/main/java/com/dalv1k/assistantbot/model/entity/Currency.java
@@ -21,6 +21,6 @@ public class Currency {
 
     private boolean tracking;
 
-    private int margin;
+    private float margin;
 
 }
